web scraping for province and taxes

+++ db/migrate/20211119042348_create_taxes.rb
@@ -1,7 +1,7 @@
 class CreateTaxes < ActiveRecord::Migration[6.1]
   def change
     create_table :taxes do |t|
-      t.integer :amount
+      t.float :amount
       t.string :province
 
       t.timestamps
